chore: move sub components to folder
