+++ src/index.js
@@ -1,2 +1,1 @@
 export { default } from './Dropdown';
-export { default as Menu } from './Menu';
