+++ test/react-dropdown-now-option.spec.js
@@ -3,21 +3,17 @@ import React from 'react';
 import sinon from 'sinon';
 import { mount } from 'enzyme';
 
-import Option from '../src/Option';
+import Option from '../src/components/Option';
 
 test('ReactDropdownNow.option, renders a string option', t => {
-  const component = mount(
-    <Option option="one" />
-  );
+  const component = mount(<Option option="one" />);
 
   t.is(component.text(), 'one');
   t.false(component.find('.Dropdown-option').hasClass('is-selected'));
 });
 
 test('ReactDropdownNow.option, renders a string option, selected', t => {
-  const component = mount(
-    <Option option="one" selected="one" />
-  );
+  const component = mount(<Option option="one" selected="one" />);
 
   t.is(component.text(), 'one');
   t.true(component.find('.Dropdown-option').hasClass('is-selected'));
@@ -25,7 +21,7 @@ test('ReactDropdownNow.option, renders a string option, selected', t => {
 
 test('ReactDropdownNow.option, renders an object option', t => {
   const component = mount(
-    <Option option={{ label: 'label', value: 'value' }} />
+    <Option option={{ label: 'label', value: 'value' }} />,
   );
 
   t.is(component.find('.Dropdown-option').text(), 'label');
@@ -34,7 +30,7 @@ test('ReactDropdownNow.option, renders an object option', t => {
 
 test('ReactDropdownNow.option, renders an object option, selected', t => {
   const component = mount(
-    <Option option={{ label: 'label', value: 'value' }} selected="value" />
+    <Option option={{ label: 'label', value: 'value' }} selected="value" />,
   );
 
   t.is(component.find('.Dropdown-option').text(), 'label');
@@ -43,17 +39,13 @@ test('ReactDropdownNow.option, renders an object option, selected', t => {
 
 test('ReactDropdownNow.option, emits onSelect event', t => {
   const onSelect = sinon.spy();
-  const component = mount(
-    <Option option="option" onSelect={onSelect} />
-  );
+  const component = mount(<Option option="option" onSelect={onSelect} />);
 
-  component
-    .find('.Dropdown-option')
-    .simulate('mousedown', { button: 0 });
+  component.find('.Dropdown-option').simulate('mousedown', { button: 0 });
 
   t.true(onSelect.calledOnce);
 
-  const [ event, value, label ] = onSelect.args[0];
+  const [event, value, label] = onSelect.args[0];
 
   t.is(event.type, 'mousedown');
   t.is(value, 'option');
